% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art_pals.R
\name{art_pals}
\alias{art_pals}
\title{Art Pals}
\usage{
art_pals(pal = NULL, n = 5)
}
\arguments{
\item{pal}{A character vector of the desired aRttools palette type.}

\item{n}{The numbers of colors desired in the output.}
}
\value{
A character vector.
}
\description{
Art Pals
}
\examples{
icecream_pal <- art_pals("icecream", 20)

}
