% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artpack-package.R
\docType{package}
\name{artpack-package}
\alias{artpack}
\alias{artpack-package}
\title{artpack: Creates Generative Art}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functionality for creating data for generative art.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://meghansaha.github.io/artpack/}
  \item Report bugs at \url{https://github.com/Meghansaha/artpack/issues}
}

}
\author{
\strong{Maintainer}: Meghan Harris \email{meghanha01@gmail.com} (\href{https://orcid.org/0000-0003-3922-8101}{ORCID}) [copyright holder]

}
\keyword{internal}
