% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_maker.R
\name{grid_maker}
\alias{grid_maker}
\title{Create Data for A Custom-built Square Grid}
\usage{
grid_maker(xlim, ylim, size, pal = NULL, fill_style = "range")
}
\arguments{
\item{xlim}{A numeric vector with two X limits. A minimum and maximum limit for the X axis.}

\item{ylim}{A numeric vector with two Y limits. A minimum and maximum limit for the Y axis.}

\item{size}{A numeric input. The size of the grid. How many shapes will appear in a single row or column.}

\item{pal}{Optional. A character vector of color codes to be applied to the grid.}

\item{fill_style}{Optional. A character input. "range" or "random". Determines how the color palette is mapped.}
}
\value{
a tibble
}
\description{
Creates a dataframe of \code{x} and \code{y} points to visualize a square grid based on given \code{x} and \code{y} limits.
Providing a color palette and fill style are optional.
}
\examples{
library(ggplot2)
grid_data <- grid_maker(
  xlim = c(0, 1),
  ylim = c(0, 1),
  size = 2,
  pal = c("red", "black", "purple")
)

ggplot() +
  geom_polygon(
    data = grid_data,
    aes(x, y, group = group),
    fill = grid_data$fill
  ) +
  coord_equal()

}
