% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotator.R
\name{rotator}
\alias{rotator}
\title{Rotator
Rotates the points in a given data frame by a given angle based on a designated anchor point.}
\usage{
rotator(data = NULL, angle = 5, anchor = "center")
}
\arguments{
\item{data}{A data frame or tibble with at least \code{x} and \code{y} variables}

\item{angle}{The angle (in degrees) the points in \code{data} will be rotated around it's anchor}

\item{anchor}{The anchor point for the rotation. Default is "center"}
}
\value{
#A data frame
}
\description{
Rotator
Rotates the points in a given data frame by a given angle based on a designated anchor point.
}
\examples{
original_square <- data.frame(x = c(0,3,3,0,0),
y = c(0,0,3,3,0))
rotated_square <- rotator(data = original_square, angle = 45, anchor = "center")

ggplot2::ggplot()+
  ggplot2::geom_path(data = original_square, ggplot2::aes(x,y), color = "red")+
  ggplot2::geom_polygon(data = rotated_square, ggplot2::aes(x,y), fill = "purple")+
  ggplot2::coord_equal()

}
