% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_data.R
\name{square_data}
\alias{square_data}
\title{Data Generation for Squares}
\usage{
square_data(x, y, size, group_var = FALSE)
}
\arguments{
\item{x}{Numeric - The bottom left \code{x} value of the square.}

\item{y}{Numeric - The bottom left \code{y} value of the square.}

\item{size}{Numeric - The size of the square. Must be a value greater than zero.}

\item{group_var}{\code{TRUE}/\code{FALSE}. Default is \code{FALSE}. If \code{TRUE}, adds a grouping variable to the data frame.
Default is switched to \code{TRUE} when more than one \code{x}, \code{y}, and \code{size} values are present.}
}
\value{
A Tibble
}
\description{
A tool for creating a data frame of values that create a square with a specified size
when plotted.
}
\examples{
one_square <- square_data(x = 0, y = 0, size = 5)

one_square |>
ggplot(aes(x,y))+
geom_path(color = "green")+
coord_equal()

x_vals <- c(0,4)
y_vals <- c(0,0)
sizes <- c(1,3)
sq_cols <- c("purple", "yellow")

two_squares <- square_data(x = x_vals, y = y_vals, size = sizes)

two_squares |>
ggplot(aes(x,y, group = group, fill = group))+
scale_fill_manual(values = sq_cols)+
theme(legend.position = "none")+
geom_polygon()+
coord_equal()


}
