% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artpack-package.R
\docType{package}
\name{artpack-package}
\alias{artpack}
\alias{artpack-package}
\title{artpack: "Assists with Creating Generative Art"}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

aRtpack is a package that contains a tool set for aRtists of all skill levels. aRtpack seeks to make life easier for more seasoned aRtists by providing helper functions to save time on more complex pieces. aRtpack also seeks to introduce more beginners into Rtistry by providing quick and easy ways to create aRt right away.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://meghansaha.github.io/artpack/}
}

}
\author{
\strong{Maintainer}: Meghan Harris \email{meghanha01@gmail.com} (\href{https://orcid.org/0000-0003-3922-8101}{ORCID}) [copyright holder]

}
\keyword{internal}
